Propagate integer overflow fixes into c_code

+++ collision_count/c_code/collision_count_sequential_linear.h
@@ -12,7 +12,7 @@ void test_count(int3 *vector, int size, int iters){
 	int beg = clock();
 	
 	int axisSize = (size + 1) * 2;
-	unsigned long memSize = sizeof(AXISTYPE) * axisSize * axisSize * axisSize;
+	unsigned long memSize = sizeof(AXISTYPE) * axisSize * axisSize * (unsigned long int) axisSize;
 	AXISTYPE *space3d = (AXISTYPE *) malloc(memSize);
 	printf("Memsize: %lf Gb, Pointer: %p\n", memSize / (double) 1024 / 1024 / 1024, space3d);
 	for(i = 0; i < iters; i++){
